PR refactoring.
