--- web/src/components/PublishProcessPage.vue
@@ -1,25 +1,0 @@
-<!-- Copyright (C) 2023 by Posit Software, PBC. -->
-
-<template>
-  <q-breadcrumbs>
-    <q-breadcrumbs-el
-      label="Project"
-      :to="{
-        name:
-          'project'
-      }"
-    />
-    <q-breadcrumbs-el
-      label="deployment url goes here"
-    />
-    <q-breadcrumbs-el
-      label="Publish Process"
-    />
-  </q-breadcrumbs>
-
-  <h1>Publish Process Page</h1>
-</template>
-
-<script setup lang="ts">
-</script>
-
