+++ web/src/stores/events.ts
@@ -326,7 +326,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.currentStep = 'createNewDeployment';
       publishStatus.steps.createNewDeployment.completion = 'inProgress';
-      publishStatus.steps.createNewDeployment.logs.push(msg);
+      publishStatus.steps.createNewDeployment.allMsgs.push(msg);
     }
   };
 
@@ -337,7 +337,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.createNewDeployment.completion = 'success';
-      publishStatus.steps.createNewDeployment.logs.push(msg);
+      publishStatus.steps.createNewDeployment.allMsgs.push(msg);
     }
   };
 
@@ -349,7 +349,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.createNewDeployment.completion = 'error';
       publishStatus.steps.createNewDeployment.error = splitMsgIntoKeyValuePairs(msg.data);
-      publishStatus.steps.createNewDeployment.logs.push(msg);
+      publishStatus.steps.createNewDeployment.allMsgs.push(msg);
     }
   };
 
@@ -361,7 +361,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.currentStep = 'setEnvVars';
       publishStatus.steps.setEnvVars.completion = 'inProgress';
-      publishStatus.steps.setEnvVars.logs.push(msg);
+      publishStatus.steps.setEnvVars.allMsgs.push(msg);
     }
   };
 
@@ -372,7 +372,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.setEnvVars.completion = 'success';
-      publishStatus.steps.setEnvVars.logs.push(msg);
+      publishStatus.steps.setEnvVars.allMsgs.push(msg);
     }
   };
   const onPublishSetEnvVarsFailure = (msg: PublishSetEnvVarsFailure) => {
@@ -383,7 +383,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.setEnvVars.completion = 'error';
       publishStatus.steps.setEnvVars.error = splitMsgIntoKeyValuePairs(msg.data);
-      publishStatus.steps.setEnvVars.logs.push(msg);
+      publishStatus.steps.setEnvVars.allMsgs.push(msg);
     }
   };
 
@@ -395,7 +395,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.currentStep = 'createBundle';
       publishStatus.steps.createBundle.completion = 'inProgress';
-      publishStatus.steps.createBundle.logs.push(msg);
+      publishStatus.steps.createBundle.allMsgs.push(msg);
     }
   };
 
@@ -406,6 +406,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.createBundle.logs.push(msg);
+      publishStatus.steps.createBundle.allMsgs.push(msg);
     }
   };
 
@@ -416,7 +417,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.createBundle.completion = 'success';
-      publishStatus.steps.createBundle.logs.push(msg);
+      publishStatus.steps.createBundle.allMsgs.push(msg);
     }
   };
 
@@ -428,7 +429,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.createBundle.completion = 'error';
       publishStatus.steps.createBundle.error = splitMsgIntoKeyValuePairs(msg.data);
-      publishStatus.steps.createBundle.logs.push(msg);
+      publishStatus.steps.createBundle.allMsgs.push(msg);
     }
   };
 
@@ -440,7 +441,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.currentStep = 'createDeployment';
       publishStatus.steps.createDeployment.completion = 'inProgress';
-      publishStatus.steps.createDeployment.logs.push(msg);
+      publishStatus.steps.createDeployment.allMsgs.push(msg);
     }
   };
 
@@ -452,7 +453,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.createDeployment.completion = 'success';
       currentPublishStatus.value.contentId = msg.data.contentId;
-      publishStatus.steps.createDeployment.logs.push(msg);
+      publishStatus.steps.createDeployment.allMsgs.push(msg);
     }
   };
 
@@ -464,7 +465,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.createDeployment.completion = 'error';
       publishStatus.steps.createDeployment.error = splitMsgIntoKeyValuePairs(msg.data);
-      publishStatus.steps.createDeployment.logs.push(msg);
+      publishStatus.steps.createDeployment.allMsgs.push(msg);
     }
   };
 
@@ -476,7 +477,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.currentStep = 'uploadBundle';
       publishStatus.steps.uploadBundle.completion = 'inProgress';
-      publishStatus.steps.uploadBundle.logs.push(msg);
+      publishStatus.steps.uploadBundle.allMsgs.push(msg);
     }
   };
 
@@ -487,7 +488,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.uploadBundle.completion = 'success';
-      publishStatus.steps.uploadBundle.logs.push(msg);
+      publishStatus.steps.uploadBundle.allMsgs.push(msg);
     }
   };
 
@@ -499,7 +500,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.uploadBundle.completion = 'error';
       publishStatus.steps.uploadBundle.error = splitMsgIntoKeyValuePairs(msg.data);
-      publishStatus.steps.uploadBundle.logs.push(msg);
+      publishStatus.steps.uploadBundle.allMsgs.push(msg);
     }
   };
 
@@ -511,7 +512,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.currentStep = 'deployBundle';
       publishStatus.steps.deployBundle.completion = 'inProgress';
-      publishStatus.steps.deployBundle.logs.push(msg);
+      publishStatus.steps.deployBundle.allMsgs.push(msg);
     }
   };
 
@@ -522,7 +523,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.deployBundle.completion = 'success';
-      publishStatus.steps.deployBundle.logs.push(msg);
+      publishStatus.steps.deployBundle.allMsgs.push(msg);
     }
   };
 
@@ -534,7 +535,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.uploadBundle.completion = 'error';
       publishStatus.steps.deployBundle.error = splitMsgIntoKeyValuePairs(msg.data);
-      publishStatus.steps.deployBundle.logs.push(msg);
+      publishStatus.steps.deployBundle.allMsgs.push(msg);
     }
   };
 
@@ -546,7 +547,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.currentStep = 'restorePythonEnv';
       publishStatus.steps.restorePythonEnv.completion = 'inProgress';
-      publishStatus.steps.restorePythonEnv.logs.push(msg);
+      publishStatus.steps.restorePythonEnv.allMsgs.push(msg);
     }
   };
 
@@ -557,7 +558,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.restorePythonEnv.logs.push(msg);
-      // Not summarizing log messages
+      publishStatus.steps.restorePythonEnv.allMsgs.push(msg);
     }
   };
 
@@ -571,7 +572,7 @@ export const useEventStore = defineStore('event', () => {
         publishStatus.steps.restorePythonEnv.progress = [];
       }
       publishStatus.steps.restorePythonEnv.progress.push(JSON.stringify(msg.data));
-      publishStatus.steps.restorePythonEnv.logs.push(msg);
+      publishStatus.steps.restorePythonEnv.allMsgs.push(msg);
     }
   };
 
@@ -585,7 +586,7 @@ export const useEventStore = defineStore('event', () => {
         publishStatus.steps.restorePythonEnv.status = [];
       }
       publishStatus.steps.restorePythonEnv.status.push(msg.data);
-      publishStatus.steps.restorePythonEnv.logs.push(msg);
+      publishStatus.steps.restorePythonEnv.allMsgs.push(msg);
       // status msg.data = {
       //     "level": "INFO",
       //     "message": "Package restore",
@@ -606,7 +607,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.restorePythonEnv.completion = 'success';
-      publishStatus.steps.restorePythonEnv.logs.push(msg);
+      publishStatus.steps.restorePythonEnv.allMsgs.push(msg);
     }
   };
 
@@ -618,7 +619,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.uploadBundle.completion = 'error';
       publishStatus.steps.restorePythonEnv.error = splitMsgIntoKeyValuePairs(msg.data);
-      publishStatus.steps.restorePythonEnv.logs.push(msg);
+      publishStatus.steps.restorePythonEnv.allMsgs.push(msg);
     }
   };
 
@@ -630,7 +631,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.currentStep = 'runContent';
       publishStatus.steps.runContent.completion = 'inProgress';
-      publishStatus.steps.runContent.logs.push(msg);
+      publishStatus.steps.runContent.allMsgs.push(msg);
     }
   };
 
@@ -641,6 +642,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.runContent.logs.push(msg);
+      publishStatus.steps.runContent.allMsgs.push(msg);
     }
   };
 
@@ -651,7 +653,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.runContent.completion = 'success';
-      publishStatus.steps.runContent.logs.push(msg);
+      publishStatus.steps.runContent.allMsgs.push(msg);
     }
   };
 
@@ -663,7 +665,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.runContent.completion = 'error';
       publishStatus.steps.runContent.error = splitMsgIntoKeyValuePairs(msg.data);
-      publishStatus.steps.runContent.logs.push(msg);
+      publishStatus.steps.runContent.allMsgs.push(msg);
     }
   };
 
@@ -675,7 +677,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.currentStep = 'setVanityURL';
       publishStatus.steps.setVanityURL.completion = 'inProgress';
-      publishStatus.steps.setVanityURL.logs.push(msg);
+      publishStatus.steps.setVanityURL.allMsgs.push(msg);
     }
   };
 
@@ -686,6 +688,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.setVanityURL.logs.push(msg);
+      publishStatus.steps.setVanityURL.allMsgs.push(msg);
     }
   };
 
@@ -696,7 +699,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.setVanityURL.completion = 'success';
-      publishStatus.steps.setVanityURL.logs.push(msg);
+      publishStatus.steps.setVanityURL.allMsgs.push(msg);
     }
   };
 
@@ -708,7 +711,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.runContent.completion = 'error';
       publishStatus.steps.setVanityURL.error = splitMsgIntoKeyValuePairs(msg.data);
-      publishStatus.steps.setVanityURL.logs.push(msg);
+      publishStatus.steps.setVanityURL.allMsgs.push(msg);
     }
   };
 
@@ -720,7 +723,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.currentStep = 'validateDeployment';
       publishStatus.steps.validateDeployment.completion = 'inProgress';
-      publishStatus.steps.validateDeployment.logs.push(msg);
+      publishStatus.steps.validateDeployment.allMsgs.push(msg);
     }
   };
 
@@ -731,6 +734,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.validateDeployment.logs.push(msg);
+      publishStatus.steps.validateDeployment.allMsgs.push(msg);
     }
   };
 
@@ -741,7 +745,7 @@ export const useEventStore = defineStore('event', () => {
     if (currentPublishStatus.value.localId === localId) {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.validateDeployment.completion = 'success';
-      publishStatus.steps.validateDeployment.logs.push(msg);
+      publishStatus.steps.validateDeployment.allMsgs.push(msg);
     }
   };
 
@@ -753,7 +757,7 @@ export const useEventStore = defineStore('event', () => {
       const publishStatus = currentPublishStatus.value.status;
       publishStatus.steps.validateDeployment.completion = 'error';
       publishStatus.steps.validateDeployment.error = splitMsgIntoKeyValuePairs(msg.data);
-      publishStatus.steps.validateDeployment.logs.push(msg);
+      publishStatus.steps.validateDeployment.allMsgs.push(msg);
     }
   };
