+++ web/src/views/existing-deployment-destination/ExistingDeploymentDestinationHeader.vue
@@ -34,10 +34,11 @@
         </div>
         <div class="row justify-left q-ma-sm q-mr-md">
           <div class="col-11">
-            <RouterLink :to="`/progress/detailed`">
-              <PublishProgressSummary
-                :id="contentId"
-              />
+            <PublishProgressSummary
+              :id="contentId"
+            />
+            <RouterLink :to="{ name: 'progress' }">
+              Log View
             </RouterLink>
           </div>
         </div>
