--- web/src/views/publish-log-view/PublishLogView.vue
@@ -1,17 +1,0 @@
-<!-- Copyright (C) 2023 by Posit Software, PBC. -->
-
-<template>
-  <q-page
-    class="max-width-md q-mx-auto"
-    padding
-  >
-    <div class="q-mt-lg">
-      <!-- <PublishSummary /> -->
-      <PublishStepper />
-    </div>
-  </q-page>
-</template>
-
-<script setup lang="ts">
-import PublishStepper from 'src/views/publish-log-view/PublishStepper.vue';
-</script>
