+++ web/src/views/publish-progress/steps/CreateDeployment.vue
@@ -14,7 +14,7 @@
 <script setup lang="ts">
 import { ref } from 'vue';
 
-import PublishStep from 'src/views/publish-log-view/PublishStep.vue';
+import PublishStep from 'src/views/publish-progress/PublishStep.vue';
 
 import { watch } from 'vue';
 import { useEventStore } from 'src/stores/events';
@@ -27,7 +27,7 @@ defineProps({
 const emit = defineEmits(['start', 'done']);
 
 const done = ref(false);
-const messages = ref(eventStore.currentPublishStatus.status.steps.createDeployment.logs);
+const messages = ref(eventStore.currentPublishStatus.status.steps.createDeployment.allMsgs);
 
 watch(
   () => eventStore.currentPublishStatus.status.steps.createDeployment.completion,
