+++ web/src/views/publish-progress/steps/RestorePythonEnvironment.vue
@@ -3,9 +3,9 @@
 <template>
   <PublishStep
     :name="name"
-    title="Upload Bundle"
-    icon="login"
-    summary="Transferring the files from your local workstation to the server."
+    title="Restore Python Environment"
+    icon="move_down"
+    summary="Installing the dependent python packages on the server in order to reproduce your runtime environment."
     :done="done"
     :messages="messages"
   />
@@ -14,7 +14,7 @@
 <script setup lang="ts">
 import { ref } from 'vue';
 
-import PublishStep from 'src/views/publish-log-view/PublishStep.vue';
+import PublishStep from 'src/views/publish-progress/PublishStep.vue';
 
 import { watch } from 'vue';
 import { useEventStore } from 'src/stores/events';
@@ -26,10 +26,10 @@ defineProps({
 const emit = defineEmits(['start', 'done']);
 
 const done = ref(false);
-const messages = ref(eventStore.currentPublishStatus.status.steps.uploadBundle.logs);
+const messages = ref(eventStore.currentPublishStatus.status.steps.restorePythonEnv.allMsgs);
 
 watch(
-  () => eventStore.currentPublishStatus.status.steps.uploadBundle.completion,
+  () => eventStore.currentPublishStatus.status.steps.restorePythonEnv.completion,
   (value) => {
     if (value === 'inProgress') {
       emit('start');
