+++ web/src/views/publish-progress/steps/index.ts
